Error handling added to User Entity

+++ src/users/users.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Body, Get, Post, Patch, Delete, Param, Query } from '@nestjs/common';
+import { Controller, Body, Get, Post, Patch, Delete, Param, Query, NotFoundException } from '@nestjs/common';
 import { createUsersDto } from './dtos/create-users.dto';
 import { UpdateUserDto } from './dtos/update-user.dto';
 import { UsersService } from './users.service';
@@ -18,17 +18,29 @@ export class UsersController {
     }
 
     @Get('/:id')
-    findOne(@Param('id') id:string){
-        return this.usersService.findUser(parseInt(id))
+    async findOne(@Param('id') id:string){
+        const user = await this.usersService.findUser(parseInt(id))
+        if(!user){
+            throw new NotFoundException('The user you are trying to find does not exists')
+        }
+        return user
     }
 
     @Delete('/delete/:id')
-    delete(@Param('id') id:string){
+    async delete(@Param('id') id:string){
+        const deletedUser = await this.usersService.findUser(parseInt(id))
+        if(!deletedUser){
+            throw new NotFoundException('The user you are trying to delete does not exists')
+        }
         return this.usersService.deleteUser(parseInt(id))
     }
 
     @Patch('/update/:id')
-    update(@Param('id') id:string, @Body() body:UpdateUserDto){
+    async update(@Param('id') id:string, @Body() body:UpdateUserDto){
+        const updatedUser = await this.usersService.findUser(parseInt(id))
+        if(!updatedUser){
+            throw new NotFoundException('The user you are trying to update does not exists')
+        }
         return this.usersService.updateUser(parseInt(id), body)
     }
 }
